Update to latest Stardust changes (#1307)
* - Updated iota.go/v3

- Renamed ExtendedOutput to BasicOutput

* - Add protocol version to basic protocol config

* - Add protocol version to info endpoint

* - Verify protocol version of messages sent over the API

- Verify the networkId in the tx essence matches the current networkId

* - Use protocol version in the gossip instead of the networkId

* - Add networkId to whiteflag checks

* - Add networkId to whiteflag confirmation checks

* - Add protocol version to messages created by coo, faucet and spammer

* - Fixing the coo in the testsuite

* - Fixed tests not compiling

* - Renamed ExtendedOutput to BasicOutput in indexer

* - Rename file to basic.go

* - Rename all things still called extended to basic

* - Added native token count to the indexer tables and removed the unused amount column

* - Added new aliasAddress filter to the indexer when filtering foundries

* - Added new hasNativeTokens, minNativeTokenCount and maxNativeTokenCount filter params to the indexer

* - Updated to latest builder pkg in iota.go/v3

* - Removed the protocol version from the config. Now using the iota.go constant

* - Fixed value integration test

* - Use uint32 for page size and native token count query params

* - Use correct unlock condition for Foundries in the Indexer

* - Added protocol version field when constructing raw messages

* - Don't gossip messages whose transaction contain a wrong networkId

* - Force protocol version when using API to send message

* - Bumping version to alpha2

* - Added the indexer plugin to all private tangle run scripts

* - Changed network to stardust-testnet-2

+++ pkg/indexer/types.go
@@ -60,7 +60,7 @@ func addressBytesForAddress(addr iotago.Address) (addressBytes, error) {
 type IndexerResult struct {
 	OutputIDs   iotago.OutputIDs
 	LedgerIndex milestone.Index
-	PageSize    int
+	PageSize    uint32
 	Cursor      *string
 	Error       error
 }
@@ -75,11 +75,11 @@ func unixTime(fromValue uint32) time.Time {
 	return time.Unix(int64(fromValue), 0)
 }
 
-func (i *Indexer) combineOutputIDFilteredQuery(query *gorm.DB, pageSize int, cursor *string) *IndexerResult {
+func (i *Indexer) combineOutputIDFilteredQuery(query *gorm.DB, pageSize uint32, cursor *string) *IndexerResult {
 
 	query = query.Select("output_id").Order("created_at asc, output_id asc")
 	if pageSize > 0 {
-		query = query.Select("output_id", "printf('%08X', strftime('%s', `created_at`)) || hex(output_id) as cursor").Limit(pageSize + 1)
+		query = query.Select("output_id", "printf('%08X', strftime('%s', `created_at`)) || hex(output_id) as cursor").Limit(int(pageSize + 1))
 
 		if cursor != nil {
 			if len(*cursor) != CursorLength {
@@ -108,7 +108,7 @@ func (i *Indexer) combineOutputIDFilteredQuery(query *gorm.DB, pageSize int, cur
 	}
 
 	var nextCursor *string
-	if pageSize > 0 && len(results) > pageSize {
+	if pageSize > 0 && uint32(len(results)) > pageSize {
 		lastResult := results[len(results)-1]
 		results = results[:len(results)-1]
 		c := strings.ToLower(lastResult.Cursor)
